feat: add indexmap feature to use ordered maps
The python configparser module uses ordered dictionaries, so that
a round-trip through a load/write keeps the same order of sections and
keys.
In Rust, the std HashMap does not keep insertion order, and actually
uses a hash that will change the order of iteration on subsequent runs.

When an INI file is to be read and/or modified by a user, it is much
preferable to keep the keys in order, as is done in python.
To do so however, a dependency is needed. `indexmap` is a popular crate
for this, that took inspiration on python's OrderedDict to create
a map tracking insertion order.

Use it in this crate, but only if a feature is activated. This keeps
the previous behavior intact, and does not bring a dependency unless
it is opted in.

The changes in themselves are very simple, as the APIs are similar.
The only change of not is in the unparse method, where the map was
cloned but most importantly, the default section was removed from the
clone. This remove messes up the order tracked by indexmap, so instead,
do not clone the map, but filter out the default section.